app/requirements.txt
django
djangorestframework
psycopg2
firebase-admin